+++ proxhy/ext/_methods.py
@@ -1,0 +1,7 @@
+methods = []
+
+
+def method(func):
+    methods.append(func)
+
+    return func
